201412311109
Originally checked in by Hujolopter\Sebastian on 2014-12-31 11:09:29.93 as Changeset 87

git-tfs-id: [https://infovision.visualstudio.com/DefaultCollection]$/RoughSets;C123

+++ RoughSets/Infovision.Datamining.Roughset.UnitTests/ReductGeneratorComparisonTest.cs
@@ -37,12 +37,12 @@ namespace Infovision.Datamining.Roughset.UnitTests
             PermutationGenerator permGenerator = new PermutationGenerator(data);
             PermutationCollection permList = permGenerator.Generate(numberOfPermutations);
             WeightGenerator weightGenerator = new WeightGeneratorConstant(data);
-            int[] epsilons;
+            double[] epsilons;
 
             //Record weights should always be used from data store. Remove weights from reduct store
             data.DataStoreInfo.RecordWeights = weightGenerator.Weights;
 
-            epsilons = new int[numberOfPermutations];
+            epsilons = new double[numberOfPermutations];
             for (int i = 0; i < numberOfPermutations; i++)
                 epsilons[i] = RandomSingleton.Random.Next(minEpsilon, maxEpsilon);
